Creacion de opciones menu
